refactor(dns): refactor dns
Full changelog
 * refactor(dns): refactor dns
 * feat(dns): start implementing reverse lookup
