--- src/dns/types.rs
@@ -1,1 +1,0 @@
-
